+++ packages/web-sdk/src/instrumentations/webVitals/instrumentation.ts
@@ -14,9 +14,9 @@ export class WebVitalsInstrumentation extends BaseInstrumentation {
   }
 
   private intializeWebVitalsInstrumentation() {
-    if (this.config.trackWebVitalsAttribution) {
-      return new WebVitalsWithAttribution(this.api.pushMeasurement, this.config.reportAllWebVitalChanges);
+    if (this.config?.trackWebVitalsAttribution) {
+      return new WebVitalsWithAttribution(this.api.pushMeasurement, this.config.webVitals);
     }
-    return new WebVitalsBasic(this.api.pushMeasurement, this.config.reportAllWebVitalChanges);
+    return new WebVitalsBasic(this.api.pushMeasurement, this.config.webVitals);
   }
 }